Organize all rpc modules in rpc directory
Rpc end-to-end test is still under the client_test directory
