+++ rpc/client.go
@@ -7,15 +7,15 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/nats-io/nats.go"
 	"github.com/rs/zerolog"
-	"github.com/statechannels/go-nitro/network"
-	"github.com/statechannels/go-nitro/network/serde"
-	"github.com/statechannels/go-nitro/network/transport"
-	natstrans "github.com/statechannels/go-nitro/network/transport/nats"
 	"github.com/statechannels/go-nitro/protocols"
 	"github.com/statechannels/go-nitro/protocols/directdefund"
 	"github.com/statechannels/go-nitro/protocols/directfund"
 	"github.com/statechannels/go-nitro/protocols/virtualdefund"
 	"github.com/statechannels/go-nitro/protocols/virtualfund"
+	"github.com/statechannels/go-nitro/rpc/network"
+	"github.com/statechannels/go-nitro/rpc/network/serde"
+	"github.com/statechannels/go-nitro/rpc/network/transport"
+	natstrans "github.com/statechannels/go-nitro/rpc/network/transport/nats"
 	"github.com/statechannels/go-nitro/types"
 
 	"github.com/statechannels/go-nitro/channel/state/outcome"
